Fix misisng get_tracr_circuits method
