+++ circuits_benchmark/commands/algorithms/acdc.py
@@ -16,9 +16,8 @@ from acdc.acdc_graphics import show
 from circuits_benchmark.benchmark.benchmark_case import BenchmarkCase
 from circuits_benchmark.commands.common_args import add_common_args
 from circuits_benchmark.training.compression.linear_compressed_tracr_transformer import LinearCompressedTracrTransformer
-from circuits_benchmark.transformers.acdc_circuit_builder import build_acdc_circuit
 from circuits_benchmark.transformers.hooked_tracr_transformer import HookedTracrTransformer
-from circuits_benchmark.utils.circuit_eval import calculate_fpr_and_tpr
+from circuits_benchmark.utils.circuit.circuit_eval import calculate_fpr_and_tpr, build_from_acdc_correspondence
 from circuits_benchmark.utils.wandb_artifact_download import download_artifact
 
 
@@ -292,15 +291,15 @@ def run_acdc(
         return_it=True,
     )
 
-    acdc_circuit = build_acdc_circuit(exp.corr)
+    acdc_circuit = build_from_acdc_correspondence(exp.corr)
     acdc_circuit.save(f"{output_dir}/final_circuit.pkl")
 
     if calculate_fpr_tpr:
         print("Calculating FPR and TPR for threshold", threshold)
         full_corr = TLACDCCorrespondence.setup_from_model(tl_model, use_pos_embed=use_pos_embed)
-        full_circuit = build_acdc_circuit(full_corr)
-        tracr_hl_circuit, tracr_ll_circuit, alignment = case.get_tracr_circuit(granularity="acdc_hooks")
-        result = calculate_fpr_and_tpr(acdc_circuit, tracr_ll_circuit, full_circuit, verbose=True)
+        full_circuit = build_from_acdc_correspondence(full_corr)
+        gt_circuit = case.get_hl_gt_circuit(granularity="acdc_hooks")
+        result = calculate_fpr_and_tpr(acdc_circuit, gt_circuit, full_circuit, verbose=True)
     else:
         result = {}
