+++ circuits_benchmark/transformers/tracr_circuits_builder.py
@@ -1,26 +1,35 @@
+from dataclasses import dataclass
 from typing import Set
 
 from networkx import DiGraph
 
 from circuits_benchmark.benchmark.vocabs import TRACR_BOS
-from circuits_benchmark.transformers.alignment import Alignment
-from circuits_benchmark.transformers.circuit import Circuit
-from circuits_benchmark.transformers.circuit_granularity import CircuitGranularity
-from circuits_benchmark.transformers.circuit_node import CircuitNode
+from circuits_benchmark.utils.circuit.alignment import Alignment
+from circuits_benchmark.utils.circuit.circuit import Circuit
+from circuits_benchmark.utils.circuit.circuit_granularity import CircuitGranularity
+from circuits_benchmark.utils.circuit.circuit_node import CircuitNode
 from tracr.compiler import nodes
 from tracr.craft.bases import VectorSpaceWithBasis
 from tracr.craft.transformers import SeriesWithResiduals, MultiAttentionHead, AttentionHead, MLP
 from tracr.rasp.rasp import Aggregate, Selector
 
+@dataclass
+class TracrCircuits:
+  tracr_variables_circuit: Circuit
+  tracr_transformer_circuit: Circuit
+  alignment: Alignment
+
 
 def build_tracr_circuits(tracr_graph: DiGraph,
                          craft_model: SeriesWithResiduals,
-                         granularity: CircuitGranularity = "component") -> (Circuit, Circuit, Alignment):
-  hl_circuit = build_hl_circuit(tracr_graph)
-  ll_circuit = Circuit(granularity)
+                         granularity: CircuitGranularity = "component") -> TracrCircuits:
+  """Returns the high-level Tracr variables circuit, the low-level circuit for Tracr-generated transformers, and the
+  alignment between both circuits."""
+  tracr_variables_circuit = build_tracr_variables_circuit(tracr_graph)
+  tracr_transformer_circuit = Circuit(granularity)
   alignment = Alignment()
 
-  if "tokens" in hl_circuit.nodes:
+  if "tokens" in tracr_variables_circuit.nodes:
     if granularity == "component":
       ll_node = "embed"
     elif granularity == "matrix":
@@ -28,10 +37,10 @@ def build_tracr_circuits(tracr_graph: DiGraph,
     else:
       ll_node = "hook_embed"
 
-    ll_circuit.add_node(CircuitNode(ll_node))
+    tracr_transformer_circuit.add_node(CircuitNode(ll_node))
     alignment.map_hl_to_ll("tokens", CircuitNode(ll_node))
 
-  if "indices" in hl_circuit.nodes:
+  if "indices" in tracr_variables_circuit.nodes:
     if granularity == "component":
       ll_node = "pos_embed"
     elif granularity == "matrix":
@@ -39,7 +48,7 @@ def build_tracr_circuits(tracr_graph: DiGraph,
     else:
       ll_node = "hook_pos_embed"
 
-    ll_circuit.add_node(CircuitNode(ll_node))
+    tracr_transformer_circuit.add_node(CircuitNode(ll_node))
     alignment.map_hl_to_ll("indices", CircuitNode(ll_node))
 
   layer = -1
@@ -52,7 +61,7 @@ def build_tracr_circuits(tracr_graph: DiGraph,
 
       for block_idx, block in enumerate(block.sub_blocks):
         if isinstance(block, AttentionHead):
-          process_attention_head_block(block, block_idx, layer, ll_circuit, granularity, alignment)
+          process_attention_head_block(block, block_idx, layer, tracr_transformer_circuit, granularity, alignment)
         else:
           raise ValueError(f"MultiAttentionHead sub-block {block_idx} is not an AttentionHead")
 
@@ -62,25 +71,25 @@ def build_tracr_circuits(tracr_graph: DiGraph,
         layer += 1
       last_component_type = "mlp"
 
-      process_mlp_block(block, layer, ll_circuit, granularity, alignment)
+      process_mlp_block(block, layer, tracr_transformer_circuit, granularity, alignment)
 
   if granularity == "acdc_hooks" or granularity == "sp_hooks":
     resid_post_node = CircuitNode(f"blocks.{layer}.hook_resid_post")
-    ll_circuit.add_node(resid_post_node)
+    tracr_transformer_circuit.add_node(resid_post_node)
 
-    hl_result_node = hl_circuit.get_result_node()
-    for ll_node in alignment.get_ll_nodes(hl_result_node, remove_predecessors_by_ll_circuit=ll_circuit):
-      ll_circuit.add_edge(ll_node, resid_post_node)
+    hl_result_node = tracr_variables_circuit.get_result_node()
+    for ll_node in alignment.get_ll_nodes(hl_result_node, remove_predecessors_by_ll_circuit=tracr_transformer_circuit):
+      tracr_transformer_circuit.add_edge(ll_node, resid_post_node)
     alignment.map_hl_to_ll(hl_result_node, resid_post_node)
 
-  # Assert that we have all the nodes and edges in the tracr_graph (which are also nodes in the hl_circuit).
+  # Assert that we have all the nodes and edges in the tracr_graph (which are also nodes in the tracr_variables_circuit).
   # Also, complete the alignment for the cases in which we have a selector or an aggregate expression.
   nodes_with_data = list(tracr_graph.nodes(data=True))
   for label, label_data in nodes_with_data:
     if nodes.MODEL_BLOCK in label_data:
       # This is label that corresponds directly to a block
       for ll_node in alignment.get_ll_nodes(label):
-        assert ll_node in ll_circuit.nodes, f"Node {ll_node} not found in circuit"
+        assert ll_node in tracr_transformer_circuit.nodes, f"Node {ll_node} not found in circuit"
     else:
       # This is a label for which we don't have a component
       expr = label_data[nodes.EXPR]
@@ -90,28 +99,28 @@ def build_tracr_circuits(tracr_graph: DiGraph,
         for other_label, other_label_data in nodes_with_data:
           other_expr = other_label_data[nodes.EXPR]
           if isinstance(other_expr, Aggregate) and other_expr.selector == expr:
-            ll_nodes = alignment.get_ll_nodes(other_label, remove_predecessors_by_ll_circuit=ll_circuit)
+            ll_nodes = alignment.get_ll_nodes(other_label, remove_predecessors_by_ll_circuit=tracr_transformer_circuit)
             assert len(ll_nodes) > 0, f"No nodes found for label {other_label}"
             for ll_node in ll_nodes:
               if granularity == "component":
                 # the same output node is the one that implements both the selector and the aggregate
-                assert ll_node in ll_circuit.nodes, f"Node {ll_node} not found in circuit"
+                assert ll_node in tracr_transformer_circuit.nodes, f"Node {ll_node} not found in circuit"
                 alignment.map_hl_to_ll(label, ll_node)
               elif granularity == "matrix":
                 # the selector is implemented by the nodes that are input for the aggregate
-                for pred_node in ll_circuit.predecessors(ll_node):
+                for pred_node in tracr_transformer_circuit.predecessors(ll_node):
                   if "W_Q" in pred_node.name or "W_K" in pred_node.name:
                     alignment.map_hl_to_ll(label, pred_node)
               elif granularity == "acdc_hooks" or granularity == "sp_hooks":
                 # the selector is implemented by the nodes that are input for the aggregate
-                for pred_node in ll_circuit.predecessors(ll_node):
+                for pred_node in tracr_transformer_circuit.predecessors(ll_node):
                   if "hook_q" in pred_node.name or "hook_k" in pred_node.name:
                     alignment.map_hl_to_ll(label, pred_node)
               else:
                 raise ValueError(f"Granularity {granularity} not supported for Select expressions")
             break
 
-  return hl_circuit, ll_circuit, alignment
+  return TracrCircuits(tracr_variables_circuit, tracr_transformer_circuit, alignment)
 
 
 def process_attention_head_block(block: AttentionHead,
@@ -264,7 +273,7 @@ def get_hl_labels_from_input_spaces(input_spaces: list[VectorSpaceWithBasis]) ->
   return list(hl_labels)
 
 
-def build_hl_circuit(tracr_graph: DiGraph) -> Circuit:
+def build_tracr_variables_circuit(tracr_graph: DiGraph) -> Circuit:
   hl_circuit = Circuit()
 
   for node in tracr_graph.nodes:
