+++ circuits_benchmark/utils/circuit/alignment.py
@@ -1,7 +1,7 @@
 from typing import Dict, Set, List
 
-from circuits_benchmark.transformers.circuit import Circuit
-from circuits_benchmark.transformers.circuit_node import CircuitNode
+from circuits_benchmark.utils.circuit.circuit import Circuit
+from circuits_benchmark.utils.circuit.circuit_node import CircuitNode
 
 
 class Alignment(object):
