+++ circuits_benchmark/utils/circuit/circuit_granularity.py
