+++ circuits_benchmark/utils/circuit/circuit_test.py
@@ -7,18 +7,18 @@ from circuits_benchmark.benchmark.cases.case_3 import Case3
 class CircuitTest(unittest.TestCase):
   def test_build_circuit_for_case_3_with_component_granularity(self):
     case = Case3()
-    hl_circuit, ll_circuit, alignment = case.get_tracr_circuit(granularity="component")
+    tracr_transformer_circuit = case.get_hl_gt_circuit(granularity="component")
 
     expected_nodes = ["embed", "pos_embed", "blocks.0.mlp", "blocks.1.attn[0]"]
     expected_edges = [("embed", "blocks.0.mlp"),
                       ("pos_embed", "blocks.1.attn[0]"),
                       ("blocks.0.mlp", "blocks.1.attn[0]")]
-    self.assertEqual(sorted([str(n) for n in ll_circuit.nodes]), sorted(expected_nodes))
-    self.assertEqual(sorted([(str(u), str(v)) for u, v in ll_circuit.edges]), sorted(expected_edges))
+    self.assertEqual(sorted([str(n) for n in tracr_transformer_circuit.nodes]), sorted(expected_nodes))
+    self.assertEqual(sorted([(str(u), str(v)) for u, v in tracr_transformer_circuit.edges]), sorted(expected_edges))
 
   def test_build_circuit_for_case_3_with_matrix_granularity(self):
     case = Case3()
-    hl_circuit, ll_circuit, alignment = case.get_tracr_circuit(granularity="matrix")
+    tracr_transformer_circuit = case.get_hl_gt_circuit(granularity="matrix")
 
     expected_nodes = ["embed.W_E", "pos_embed.W_pos",
                       "blocks.0.mlp.W_in", "blocks.0.mlp.W_out",
@@ -32,12 +32,12 @@ class CircuitTest(unittest.TestCase):
                       ("blocks.1.attn.W_K[0]", "blocks.1.attn.W_O[0]"),
                       ("blocks.1.attn.W_Q[0]", "blocks.1.attn.W_O[0]"),
                       ("blocks.1.attn.W_V[0]", "blocks.1.attn.W_O[0]")]
-    self.assertEqual(sorted([str(n) for n in ll_circuit.nodes]), sorted(expected_nodes))
-    self.assertEqual(sorted([(str(u), str(v)) for u, v in ll_circuit.edges]), sorted(expected_edges))
+    self.assertEqual(sorted([str(n) for n in tracr_transformer_circuit.nodes]), sorted(expected_nodes))
+    self.assertEqual(sorted([(str(u), str(v)) for u, v in tracr_transformer_circuit.edges]), sorted(expected_edges))
 
   def test_build_circuit_for_case_3_with_acdc_hooks_granularity(self):
     case = Case3()
-    hl_circuit, ll_circuit, alignment = case.get_tracr_circuit(granularity="acdc_hooks")
+    tracr_transformer_circuit = case.get_hl_gt_circuit(granularity="acdc_hooks")
 
     expected_nodes = ["hook_embed", "hook_pos_embed",
                       "blocks.0.hook_mlp_in", "blocks.0.hook_mlp_out",
@@ -56,12 +56,12 @@ class CircuitTest(unittest.TestCase):
                       ("blocks.1.attn.hook_k[0]", "blocks.1.attn.hook_result[0]"),
                       ("blocks.1.attn.hook_v[0]", "blocks.1.attn.hook_result[0]"),
                       ("blocks.1.attn.hook_result[0]", "blocks.1.hook_resid_post")]
-    self.assertEqual(sorted([str(n) for n in ll_circuit.nodes]), sorted(expected_nodes))
-    self.assertEqual(sorted([(str(u), str(v)) for u, v in ll_circuit.edges]), sorted(expected_edges))
+    self.assertEqual(sorted([str(n) for n in tracr_transformer_circuit.nodes]), sorted(expected_nodes))
+    self.assertEqual(sorted([(str(u), str(v)) for u, v in tracr_transformer_circuit.edges]), sorted(expected_edges))
 
   def test_build_circuit_for_case_3_with_sp_hooks_granularity(self):
     case = Case3()
-    hl_circuit, ll_circuit, alignment = case.get_tracr_circuit(granularity="sp_hooks")
+    tracr_transformer_circuit = case.get_hl_gt_circuit(granularity="sp_hooks")
 
     expected_nodes = ["hook_embed", "hook_pos_embed",
                       "blocks.0.hook_mlp_in", "blocks.0.hook_mlp_out",
@@ -76,12 +76,12 @@ class CircuitTest(unittest.TestCase):
                       ("blocks.1.hook_k_input[0]", "blocks.1.attn.hook_result[0]"),
                       ("blocks.1.hook_v_input[0]", "blocks.1.attn.hook_result[0]"),
                       ("blocks.1.attn.hook_result[0]", "blocks.1.hook_resid_post")]
-    self.assertEqual(sorted([str(n) for n in ll_circuit.nodes]), sorted(expected_nodes))
-    self.assertEqual(sorted([(str(u), str(v)) for u, v in ll_circuit.edges]), sorted(expected_edges))
+    self.assertEqual(sorted([str(n) for n in tracr_transformer_circuit.nodes]), sorted(expected_nodes))
+    self.assertEqual(sorted([(str(u), str(v)) for u, v in tracr_transformer_circuit.edges]), sorted(expected_edges))
 
   def test_build_circuit_for_case_16_with_acdc_hooks_granularity(self):
     case = Case16()
-    hl_circuit, ll_circuit, alignment = case.get_tracr_circuit(granularity="acdc_hooks")
+    tracr_transformer_circuit = case.get_hl_gt_circuit(granularity="acdc_hooks")
 
     expected_nodes = ["blocks.0.attn.hook_k[0]", "blocks.0.attn.hook_q[0]", "blocks.0.attn.hook_result[0]",
                       "blocks.0.attn.hook_v[0]", "blocks.0.hook_k_input[0]", "blocks.0.hook_mlp_in",
@@ -149,5 +149,5 @@ class CircuitTest(unittest.TestCase):
                       ("hook_pos_embed", "blocks.0.hook_mlp_in"),
                       ("hook_pos_embed", "blocks.2.hook_mlp_in"),
                       ("hook_pos_embed", "blocks.3.hook_q_input[0]")]
-    self.assertEqual(sorted([str(n) for n in ll_circuit.nodes]), sorted(expected_nodes))
-    self.assertEqual(sorted([(str(u), str(v)) for u, v in ll_circuit.edges]), sorted(expected_edges))
+    self.assertEqual(sorted([str(n) for n in tracr_transformer_circuit.nodes]), sorted(expected_nodes))
+    self.assertEqual(sorted([(str(u), str(v)) for u, v in tracr_transformer_circuit.edges]), sorted(expected_edges))
